ADOBuddyPythonVS/requirements.txt
gradio
langchain-openai
langchain-core>=0.3.25
langchain-mcp-adapters>=0.0.3
langgraph>=0.3.5
copilotkit>=0.1.39
requests
azure-devops
azure-identity
python-dotenv
typing-extensions
streamlit